+++ t/ack-color.t
@@ -8,13 +8,19 @@ use Test::More;
 use lib 't';
 use Util;
 
-plan tests => 10;
+plan tests => 12;
 
 prep_environment();
 
-my $match_start = "\e[30;43m";
-my $match_end   = "\e[0m";
-my $line_end    = "\e[0m\e[K";
+my $match_start  = "\e[30;43m";
+my $match_end    = "\e[0m";
+my $line_end     = "\e[0m\e[K";
+
+my $green_start  = "\e[1;32m";
+my $green_end    = "\e[0m";
+
+my $yellow_start = "\e[1;33m";
+my $yellow_end   = "\e[0m";
 
 NORMAL_COLOR: {
     my @files = qw( t/text/boy-named-sue.txt );
@@ -51,6 +57,44 @@ MULTIPLE_MATCHES: {
     is( $results[0], "A ${match_start}victim${match_end} of ${match_start}collision${match_end} on the open sea$line_end",
         'multiple matches highlighted' );
 }
+
+
+subtest 'Heading colors, single line' => sub {
+    plan tests => 6;
+
+    # Without the column number
+    my $file = reslash( 't/text/science-of-myth.txt' );
+    my @args = qw( mutually -w --color -H );
+    my @results = run_ack( @args, $file );
+
+    is( $results[0], "${green_start}$file${green_end}:${yellow_start}13${yellow_end}:Science and religion are not ${match_start}mutually${match_end} exclusive$line_end", 'Properly row highlights' );
+    is( scalar @results, 1, 'Only one line back' );
+
+    # With column number
+    @results = run_ack( @args, '--column', $file );
+    is( $results[0], "${green_start}$file${green_end}:${yellow_start}13${yellow_end}:${yellow_start}30${yellow_end}:Science and religion are not ${match_start}mutually${match_end} exclusive$line_end", 'Properly row highlights' );
+    is( scalar @results, 1, 'Only one line back' );
+};
+
+
+subtest 'Heading colors, grouped' => sub {
+    plan tests => 8;
+
+    # Without the column number
+    my $file = reslash( 't/text/science-of-myth.txt' );
+    my @args = qw( mutually -w --color --group );
+    my @results = run_ack( @args, 't/text' );
+
+    is( $results[0], "${green_start}$file${green_end}", 'Heading is right' );
+    is( $results[1], "${yellow_start}13${yellow_end}:Science and religion are not ${match_start}mutually${match_end} exclusive$line_end", 'Match line OK' );
+    is( scalar @results, 2, 'Exactly two lines' );
+
+    # With column number
+    @results = run_ack( @args, '--column', 't/text' );
+    is( $results[0], "${green_start}$file${green_end}", 'Heading is right' );
+    is( $results[1], "${yellow_start}13${yellow_end}:${yellow_start}30${yellow_end}:Science and religion are not ${match_start}mutually${match_end} exclusive$line_end", 'Match line OK' );
+    is( scalar @results, 2, 'Exactly two lines' );
+};
 
 done_testing();